Added in new back button to game UI so player can restart
Added in management of the objects that are DontDestroyList so they are
destroyed when a game is restarted (but still preserved between
scenes). Also made changes to tidy up code for Singleton creation.

Known bug if you try to start a local server with more than one player
and then restart for say a one player game. Probably requires updating
cancel buttons to behave like the new back button in the game UI.